[sensorlib] Fix down sampling sample size issue (#168)
# PR Details
## Description

This PR fixes a bug with the `SemanticLidarSensor`'s down sampling. If a
detected object's hit point population was smaller than the down
sampling sample size, the code would crash. To resolve the issue, the
sampling size is considered a maximum sample size, and populations
smaller than the size will remain unchanged.

## Related GitHub Issue

Closes #166

## Related Jira Key

Closes [CDAR-476](https://usdot-carma.atlassian.net/browse/CDAR-476)

## Motivation and Context

The `sensorlib` library would crash when a `SemanticLidarSensor` has a
sparse hit point set for a detected object.

## How Has This Been Tested?

Manually verified that proposed change resolves the issue.

## Types of changes

- [x] Defect fix (non-breaking change that fixes an issue)

## Checklist:

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.


[CDAR-476]:
https://usdot-carma.atlassian.net/browse/CDAR-476?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ